Set server idle timeout higher than Go's default client timeout
We're seeing a race condition where a Go client with connection
Keep-Alive is having connections reset because the 5 second header
timeout is closing the connection before receiving an in-flight
request. The default idle timeout for a Go HTTP client is 90s, so
a 120s server side timeout will ensure the client is closing the
connection rather than the server.

At some point, we should make this configurable.