Middleware / Missing handler robustness
Previously, middleware did not run if the specified method was missing,
leaving no way to implement custom handling for missing methods.

This restructures how middleware are handled so that the middleware
run regardless of the handler.

In addition to that, it's now possible to register a missing handler,
so that applcations can implement custom handlers for when they don't
implement a method.