Rejiggering "latest" unmarshalling
The old version of "latest" unmarshalling had a problem:

If the header passed for "latest" was too far in the future,
applications would end up reporting "Block Not Found" when
the user asked for the latest block. Asking for "latest"
is essentially saying "I don't care enough to know and
specify what "latest" means, I just want something current."
To have that come back with "Block Not Found" reasonably
violates user expectations, even if it should be very rare.

So we decided we want to await the latest block number and
only pass a real block number to the application if the height
feed says that number is available. But there were some challenges:

* That meant we needed to await before assigning a value to latest
* But we don't want to await on RPC calls that don't ask for latest
* Awaiting while holding the marshal lock would delay every user of
  the system while we did the resolution, which would be very bad.
* We still want to await specified user-specified block numbers when
  they didn't specify latest

This solution solves those problems, with one caveat. We track pointers
for all the unmarshalled "latest" values. If we have any such pointers,
as soon as we've unlocked the unmarshaller we await the block indicated
by the header. If await indicates that the block resolved successfully,
we assign that value to all the pointers in the list.

The caveat is that previously, consumers like Cardinal and Flume could
refer to rpc.LatestBlock during the unmarshalling process and have it
resolve to the header-specified value. This is no longer true, and those
tools will need to explicitly unmarshall BlockNumbers with the JSON
library to ensure they get handled properly.