Add support for safe/finalized
This extends the height feed to take a new HeightRecord, which can
include safe/finalized information, and "safe" or "finalized" values
in RPCBlockNumber types will JSON unmarshal to the latest safe /
finalized value on the height feed.

The new heightfeed is registered through a different method than the
old one to keep backwards compatibility for the time being.